backend/pyproject.toml
[tool.poetry]
name = "rsatlantis"
version = "0.1.0"
description = ""
authors = ["John Victor <johnvictorfs@gmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10.3"
django-cors-headers = "^3.1.1"
django-rest-auth = "^0.9.5"
whitenoise = "^4.1"
bleach = "^3.1"
django-debug-toolbar = "^1.11"
requests = "^2.22"
django-extensions = "^2.1"
django-dynamic-preferences = "^1.7"
gunicorn = "^19.9"
psycopg2-binary = "^2.7"
markdown = "^3.1"
drf-yasg = "^1.17.0"
django-webpack-loader = "^0.6.0"

[tool.poetry.dev-dependencies]
ipython = "^7.2"
jupyter = "^1.0"
coverage = "^4.5"
flake8 = "^3.7"
mypy = "^0.750"
flake8-bugbear = "^19.8.0"
autopep8 = "^1.4.4"
flake8-quotes = "^2.1.1"
isort = "^4.3.21"
django-stubs = "^1.3.2"
pre-commit = "^1.20.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
